% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{run_multiple_osmium}
\alias{run_multiple_osmium}
\title{Runs osmium}
\usage{
run_multiple_osmium(pmids, density_alpha = 1)
}
\arguments{
\item{pmids}{A character vector of PMIDs}

\item{density_alpha}{A density_alpha to modulate the weight of pages}
}
\description{
Runs osmium
}
